+++ src/QuickMarcEditor/constants.js
@@ -11,30 +11,6 @@ export const LEADER_EDITABLE_BYTES = {
   [MARC_TYPES.AUTHORITY]: [5, 17, 18],
 };
 
-export const NON_BREAKING_SPACE = '\u00A0';
-
-export const LEADER_VALUES_FOR_POSITION = {
-  [MARC_TYPES.BIB]: {
-    5: ['a', 'c', 'd', 'n', 'p'],
-    6: ['a', 'c', 'd', 'e', 'f', 'g', 'i', 'j', 'k', 'm', 'o', 'p', 'r', 't'],
-    7: ['a', 'b', 'c', 'd', 'i', 'm', 's'],
-    8: ['\\', ' ', NON_BREAKING_SPACE, 'a'],
-    18: ['\\', ' ', NON_BREAKING_SPACE, 'a', 'c', 'i', 'n', 'u'],
-    19: ['\\', ' ', NON_BREAKING_SPACE, 'a', 'b', 'c'],
-  },
-  [MARC_TYPES.HOLDINGS]: {
-    5: ['c', 'd', 'n'],
-    6: ['u', 'v', 'x', 'y'],
-    17: ['1', '2', '3', '4', '5', 'm', 'u', 'z'],
-    18: ['\\', ' ', NON_BREAKING_SPACE, 'i', 'n'],
-  },
-  [MARC_TYPES.AUTHORITY]: {
-    5: ['a', 'c', 'd', 'n', 'o', 's', 'x'],
-    17: ['n', 'o'],
-    18: ['\\', ' ', NON_BREAKING_SPACE, 'c', 'i', 'u'],
-  },
-};
-
 export const LEADER_DOCUMENTATION_LINKS = {
   [MARC_TYPES.BIB]: 'https://loc.gov/marc/bibliographic/bdleader.html',
   [MARC_TYPES.HOLDINGS]: 'https://www.loc.gov/marc/holdings/hdleader.html',
@@ -82,12 +58,6 @@ export const FIELDS_TAGS_WITHOUT_DEFAULT_SUBFIELDS = {
     { tag: '008' },
   ],
 };
-
-export const CREATE_BIB_RECORD_DEFAULT_LEADER_VALUE = '00000n\\\\\\a2200000uu\\4500';
-
-export const CREATE_HOLDINGS_RECORD_DEFAULT_LEADER_VALUE = '00000nu\\\\\\2200000un\\4500';
-
-export const CREATE_AUTHORITY_RECORD_DEFAULT_LEADER_VALUE = '00000nz\\\\a2200000o\\\\4500';
 
 export const CREATE_HOLDINGS_RECORD_DEFAULT_FIELD_TAGS = ['001', '004', '005', '008', '852', '999'];